Call of duty bracket made for a tournament
This is an Excel file that was originally made in Google Sheets and converted. The file has multiple functions running off the scoreboard that automatically calculate the total points someone has accrued from kills and placement in the tournament. There are graphs at the bottom of the main page that pull from the scoreboard showing all kills and placements compared to other competitors. This is meant to track overall performance. On the other pages, the graphs pull from the scoreboard to show personal kills and placement per game to track performance over time.