First attempt at tournament bracket
This was my very first attempt at making a call of duty bracket and was a big inspiration to make the other one in this repository. This one is much more bare-bones but follows a similar points system. The main difference between the two as far as functions go is this one has team placement while the other does not.