Mario Kart bracket using a scripted button
This is another tournament bracket I made that uses a button to automatically sort competitors into heats based on skill level. Although simple, this was my first attempt at using scripts attached to objects in my Google Sheets. The way it works is the competitors will enter their name and their self-proclaimed skill level. Once all competitors have entered their information, the button can be pressed and the script will automatically sort the competitors from highest skill to lowest skill and split them into heats of 4. The heat tabs automatically get populated with competitors after the button press. In each heat tab, the points are automatically calculated using a switch case which is another new thing in this sheet. The final heat tab will be automatically populated with the current winners of each heat. As a result of this being created in google sheets and using a script on the button, this sheet will only work within google sheets.